:bug: [#149955581] Switch to original dep. solver
For the new versions of `opkg`, `libsolv` is used as the default
dependency solving library instead of the one built by opkg itself.
`opkg` using `libsolv`, however, does not support the use of the
`list-upgradable` command. Therefore the use of `libsolv` was disabled
until the use of this command is fixed in the future.