[#153210814] Work around the halting bug (#9)
* :bug::hankey: [#153210814] Work around the halting bug

By not stopping the kernel after a shutdown, the watchdog will not
intervene. However the kernel keeps running, the filesystem is already
unmounted, so there should be no nasty side effects. No state is saved
that could break the system.

* :rewind: [#153210814] Remove unintended change

The `-p` argument was removed unintentionally.

* :memo: [#153210814] Add explanatory commentary 

The choice was made to remove a piece of code, but it is not really clear why. 
Therefore some commentary was added for future reference.