➕ [#160509631] Add `ar` command to busybox
The `updater` package really only has one file inside it, which is the
`checkUpdates` script. This script is used to download, check and then
install updates using opkg. For checking packages, the `ar` command is
used. If `ar` fails, normally the same thing is tried using `opkg` and
if this fails, the file check fails. However, at the beginning of this
update script, a sanity check is done to see if `ar` is found. When it
is not found, the script stops. So if the system does not have an `ar`
command available, it is also unable to install it afterwards. This is
really not wat we want.