⬆ [#154624211] Upgrade opkg version to 0.4.0
We see a problem with opkg version 0.3.4. It takes several of minutes
between downloading the packages and performing the upgrade. This was
introduced somewhere between versions 0.2.3 and 0.3.4, so in order to
fix this problem we could downgrade to 0.2.3 or upgrade to >0.3.6. We
chose to do the latter because it is always better to be up-to-date.

Also, with this newer version of libsolv, the command list-upgradable
is implemented so we can finally make use of this vastly superior way
of resolving dependencies.