👆🏻 [#163979845] Upgrade openssl from 1.0.2j to 1.1.1a
Usually we want to be reluctand in updating packages when it is not
strictly necessary, but the ca-certificates package uses an openssl
command called `openssl rehash`. In the old situation an equivalent
script existed called `c_rehash`. The ca-certificates packages uses
`rehash` in the post-install. This command only exists in the newer
versions of openssl. The other possibility would have been to alter
the postinstall script to use `c_rehash` instead of `rehash`. It is
however always better to update when taking security into account.