💉 [#164163496] Fix update-ca-certificates postinstall
The CA-certificates package runs the command `update-ca-certificates`
in the post install script. This command actually is a script itself.
It makes use of the `openssl rehash` command. This command however is
not in the version of openssl we use. We use version 1.0.2j, while it
is introduced in 1.1.0. Updating openssl is not doable for the moment
because we want to release quickly and do not have the time to fit in
openssl 1.1.0+ because it is widely used throughout the yocto project
and poky. We can always choose to update it in the future.

Essentially, this patch reverts the following commit:

    1bc87e0b41a04551a93d4e784e158b044c18792a